Totosteps/requirements.txt
authlib ~= 1.2
django ~= 4.2
python-dotenv ~= 1.0
requests ~= 2.31
djangorestframework==3.15.2
python-dateutil==2.9.0
Pillow==10.4.0
djangorestframework-simplejwt==5.3.1 
pyjwt==2.9.0